pyproject.toml
[tool.poetry]
name = "lesson-2"
version = "0.1.0"
description = ""
authors = ["Allen8Kane <allen8kane@gmail.com>"]
readme = "README.md"
packages = [{include = "lesson_2"}]

[tool.poetry.dependencies]
python = "^3.9.5"
djangorestframework = "^3.14.0"
django = "^4.1.7"
django-cors-headers = "^3.14.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
